Solved AoC2021 Day 14 Part 1

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day14/Day14Test.kt
@@ -4,13 +4,14 @@ import de.havox_design.aoc.utils.kotlin.helpers.tests.shouldBe
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
+import java.math.BigInteger
 import java.util.stream.Stream
 
 class Day14Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: BigInteger) =
         ExtendedPolymerization(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
@@ -22,7 +23,7 @@ class Day14Test {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day14/day14part1sample.txt", 1588L)
+                Arguments.of("de/havox_design/aoc2021/day14/day14part1sample.txt", BigInteger.valueOf(1588))
             )
 
         @JvmStatic
